Delete like arreglado

+++ app/Http/Controllers/LikesController.php
@@ -62,10 +62,17 @@ class LikesController extends Controller
 
     public function delete(likes $likes, $id)
     {
-        $Likes = likes::findOrFail($id);
-        $Likes->delete(); 
+       
+        $Likes = likes::where("id_user", $id_user)
+                                ->where("id_interest", $id_interest)
+                                ->first();
 
-        return ["response" => "Object with ID $id Deleted"];
+        if ($Likes){
+        $Likes->delete(); 
+        return ["response" => "Object Deleted"];
+        }
+
+        return response()->json(['response' => 'Object not found'], 404);
         
     }
